+++ packages/package-json-lint-config-terra/package-json-lint.config.js
@@ -3,5 +3,6 @@ module.exports = {
     'require-no-hard-coded-dependency-versions': { severityType: 'error' },
     'require-no-terra-base-peer-dependency-versions': { severityType: 'warn' },
     'require-theme-context-versions': { severityType: 'warn' },
+    'require-ie10-compatible-dependency-versions': { severityType: 'warn' },
   },
 };