+++ packages/package-json-lint-config-terra/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added new rule `require-ie10-compatible-dependency-versions`.
+
 ## 1.1.0 - (September 28, 2021)
 
 * Changed