+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation for rule `require-ie10-compatible-dependency-versions`.
+
 ## 2.5.0 - (September 28, 2021)
 
 * Added